Makefile: Remove non--object file targets' dependencies on the toolchain installation check
This offers the advantage of allowing one to copy the Makefile into a
subdirectory (e.g. riscv-tests/isa/rv32ui) and build files based on preexisting
object files without forcing the latter to be rebuilt.  In the case of the unit
tests, they are difficult to build manually, but with this commit, you can run
 $ ./unittest testname
from the root directory once, copy the Makefile, then
 $ make testname.whatever
from the subdirectory to test the build steps.